[#35] converted the gui script file to have it's own class: AsimovOperation. few minor silent bugs detected while clicking on control buttons
